+++ docs/Changelog.md
@@ -10,6 +10,8 @@
   scenarios with multiple sources can be tested from file
 * [#3869](https://github.com/TouK/nussknacker/pull/3869) cross-compile - scala 2.12 & 2.13
 * [#3874](https://github.com/TouK/nussknacker/pull/3874) Tumbling window with OnEvent trigger saves context
+* [#3853](https://github.com/TouK/nussknacker/pull/3853) Support of patternProperties in sink with JSON Schema
+
 
 1.7.0
 ------------------------